requirements.txt
click==6.7
Cython==0.25.2
Flask==0.12
funcsigs==1.0.2
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.6.0
h5py==2.6.0
itsdangerous==0.24
Jinja2==2.9.4
Keras==1.2.2
MarkupSafe==0.23
mock==2.0.0
numpy==1.12.0
olefile==0.44
pbr==1.10.0
Pillow==4.0.0
protobuf==3.1.0.post1
PyYAML==3.12
scipy==0.18.1
six==1.10.0
tensorflow==1.0.0
Theano==0.8.2
Werkzeug==0.11.15
